feat: implement base for tile parsing

+++ src/entityManager.cc
@@ -34,7 +34,7 @@ std::shared_ptr<AbstractGameObject> EntityManager::getEntityByTag(OBJECT_TAG tag
     }
   }
 
-  printf("ERROR: Tried getting an entity by a tag that doesn't exist. Tag: %d\n", tag);
+  printf("ERROR: Tried getting an entity by a tag that doesn't exist. Tag: %d\n", (int)tag);
   return nullptr;
 }
